+++ src/app/overlay/overlay-archive-delete-filter-mobile/overlay-archive-delete-filter-mobile.component.ts
@@ -79,7 +79,6 @@ export class OverlayArchiveDeleteFilterMobileComponent implements OnInit {
   }
 
   onLoad(): void {
-    console.log('df');
     this.loading = false;
   }